Refactor code location
